pyproject.toml
[project]
name = "daily-paper-rss-ai-enhance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.0.0",
    "crawl4ai>=0.7.7",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "langchain>=1.1.0",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "pyzotero>=1.7.5",
    "schedule>=1.2.2",
    "tavily-python>=0.7.14",
    "tqdm>=4.67.1",
]
